refactor: make naming more consistent between credentials, versions and locations modules
repositories are interfaces to be implemented by the server service provider

services are the classes implemented by the library, relying on repositories
