+++ ocpi-toolkit-2.2.1/src/main/kotlin/com/izivia/ocpi/toolkit/modules/locations/services/LocationsCpoService.kt
@@ -1,52 +1,81 @@
 package com.izivia.ocpi.toolkit.modules.locations.services
 
+import com.izivia.ocpi.toolkit.common.CiString
+import com.izivia.ocpi.toolkit.common.OcpiResponseBody
 import com.izivia.ocpi.toolkit.common.SearchResult
+import com.izivia.ocpi.toolkit.common.validation.validate
+import com.izivia.ocpi.toolkit.common.validation.validateDates
+import com.izivia.ocpi.toolkit.common.validation.validateInt
+import com.izivia.ocpi.toolkit.common.validation.validateLength
+import com.izivia.ocpi.toolkit.modules.locations.LocationsCpoInterface
 import com.izivia.ocpi.toolkit.modules.locations.domain.Connector
 import com.izivia.ocpi.toolkit.modules.locations.domain.Evse
 import com.izivia.ocpi.toolkit.modules.locations.domain.Location
+import com.izivia.ocpi.toolkit.modules.locations.repositories.LocationsCpoRepository
 import java.time.Instant
 
-/**
- * - GET: Fetch a list locations, last updated between the {date_from} and {date_to} (paginated), or get a specific
- * location, EVSE or Connector.
- * - POST: n/a
- * - PUT: n/a
- * - PATCH: n/a
- * - DELETE: n/a
- */
-interface LocationsCpoService {
-    /**
-     * If additional parameters: {date_from} and/or {date_to} are provided, only Locations with
-     * (last_updated) between the given date_from and date_to will be returned. If an EVSE is
-     * updated, also the 'parent' Location's last_updated fields is updated. If a Connector is
-     * updated, the EVSE's last_updated and the Location's last_updated field are updated.
-     *
-     * This request is paginated, it supports the pagination related URL parameters.
-     *
-     * @param dateFrom Instant? Only return Locations that have last_updated after this Date/Time.
-     * @param dateTo Instant? Only return Locations that have last_updated before this Date/Time.
-     * @param offset Int? The offset of the first object returned. Default is 0.
-     * @param limit Int? Maximum number of objects to GET.
-     * @return List<Location> The endpoint returns a list of Location objects The header will
-     * contain the pagination related headers.
-     */
-    fun getLocations(dateFrom: Instant?, dateTo: Instant?, offset: Int = 0, limit: Int?): SearchResult<Location>
+class LocationsCpoService(
+    private val service: LocationsCpoRepository
+) : LocationsCpoInterface {
 
-    /**
-     * @param locationId String max-length = 39
-     */
-    fun getLocation(locationId: String): Location?
+    override fun getLocations(
+        dateFrom: Instant?,
+        dateTo: Instant?,
+        offset: Int,
+        limit: Int?
+    ): OcpiResponseBody<SearchResult<Location>> = OcpiResponseBody.of {
+        validate {
+            if (dateFrom != null && dateTo != null) validateDates("dateFrom", dateFrom, "dateTo", dateTo)
+            if (limit != null) validateInt("limit", limit, 0, null)
+            validateInt("offset", offset, 0, null)
+        }
 
-    /**
-     * @param locationId String max-length = 39
-     * @param evseUid String? max-length = 39
-     */
-    fun getEvse(locationId: String, evseUid: String): Evse?
+        service
+            .getLocations(dateFrom, dateTo, offset, limit)
+            .also { searchResult ->
+                searchResult.list.forEach { location -> location.validate() }
+            }
+    }
 
-    /**
-     * @param locationId String max-length = 39
-     * @param evseUid max-length = 39
-     * @param connectorId max-length = 39
-     */
-    fun getConnector(locationId: String, evseUid: String, connectorId: String): Connector?
+    override fun getLocation(
+        locationId: CiString
+    ): OcpiResponseBody<Location?> = OcpiResponseBody.of {
+        validate {
+            validateLength("locationId", locationId, 36)
+        }
+
+        service
+            .getLocation(locationId)
+            ?.validate()
+    }
+
+    override fun getEvse(
+        locationId: CiString,
+        evseUid: CiString
+    ): OcpiResponseBody<Evse?> = OcpiResponseBody.of {
+        validate {
+            validateLength("locationId", locationId, 36)
+            validateLength("evseUid", evseUid, 36)
+        }
+
+        service
+            .getEvse(locationId, evseUid)
+            ?.validate()
+    }
+
+    override fun getConnector(
+        locationId: CiString,
+        evseUid: CiString,
+        connectorId: CiString
+    ): OcpiResponseBody<Connector?> = OcpiResponseBody.of {
+        validate {
+            validateLength("locationId", locationId, 36)
+            validateLength("evseUid", evseUid, 36)
+            validateLength("connectorId", connectorId, 36)
+        }
+
+        service
+            .getConnector(locationId, evseUid, connectorId)
+            ?.validate()
+    }
 }
