+++ ocpi-toolkit-2.2.1/src/main/kotlin/com/izivia/ocpi/toolkit/modules/versions/services/VersionsService.kt
@@ -1,11 +1,11 @@
-package com.izivia.ocpi.toolkit.modules.versions.validation
+package com.izivia.ocpi.toolkit.modules.versions.services
 
 import com.izivia.ocpi.toolkit.common.OcpiResponseBody
 import com.izivia.ocpi.toolkit.modules.versions.VersionsInterface
 import com.izivia.ocpi.toolkit.modules.versions.domain.Version
 import com.izivia.ocpi.toolkit.modules.versions.repositories.VersionsRepository
 
-class VersionsValidationService(
+class VersionsService(
     private val repository: VersionsRepository
 ) : VersionsInterface {
