+++ ocpi-toolkit-2.2.1/src/test/kotlin/com/izivia/ocpi/toolkit/tests/integration/mock/LocationsCpoMongoRepository.kt
@@ -5,16 +5,16 @@ import com.izivia.ocpi.toolkit.common.toSearchResult
 import com.izivia.ocpi.toolkit.modules.locations.domain.Connector
 import com.izivia.ocpi.toolkit.modules.locations.domain.Evse
 import com.izivia.ocpi.toolkit.modules.locations.domain.Location
-import com.izivia.ocpi.toolkit.modules.locations.services.LocationsCpoService
+import com.izivia.ocpi.toolkit.modules.locations.repositories.LocationsCpoRepository
 import com.mongodb.client.MongoCollection
 import org.litote.kmongo.and
 import org.litote.kmongo.gte
 import org.litote.kmongo.lte
 import java.time.Instant
 
-class LocationsCpoMongoService(
+class LocationsCpoMongoRepository(
     private val collection: MongoCollection<Location>
-) : LocationsCpoService {
+) : LocationsCpoRepository {
 
     override fun getLocations(dateFrom: Instant?, dateTo: Instant?, offset: Int, limit: Int?): SearchResult<Location> =
         collection
