+++ ocpi-toolkit-2.2.1/src/test/kotlin/com/izivia/ocpi/toolkit/tests/mock/LocationsEmspRepository.kt
@@ -1,13 +1,13 @@
 package com.izivia.ocpi.toolkit.tests.mock
 
 import com.izivia.ocpi.toolkit.modules.locations.domain.*
-import com.izivia.ocpi.toolkit.modules.locations.services.LocationsEmspService
+import com.izivia.ocpi.toolkit.modules.locations.repositories.LocationsEmspRepository
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.slot
 import java.util.*
 
-fun locationsEmspService(locations: List<Location>): LocationsEmspService = mockk {
+fun locationsEmspService(locations: List<Location>): LocationsEmspRepository = mockk {
     val countryCode = slot<String>()
     val partyId = slot<String>()
     val locationId = slot<String>()
