+++ ocpi-toolkit-2.2.1/src/test/kotlin/com/izivia/ocpi/toolkit/tests/validation/LocationsValidatorTest.kt
@@ -1,7 +1,7 @@
 package com.izivia.ocpi.toolkit.tests.validation
 
 import com.izivia.ocpi.toolkit.modules.locations.domain.*
-import com.izivia.ocpi.toolkit.modules.locations.validation.validate
+import com.izivia.ocpi.toolkit.modules.locations.services.validate
 import com.izivia.ocpi.toolkit.samples.common.*
 import org.junit.jupiter.api.Test
 import strikt.api.expectCatching
